requirements.txt
numpy
ipykernel
pandas
scikit-learn
pyyaml
pymongo
from_root
matplotlib
imblearn
python-dotenv
certifi
requests
pytest
boto3
fastapi
uvicorn
seaborn
-e .